Nightly commit

+++ firmware/src/app_sdcard_write.h
@@ -53,7 +53,7 @@ void APP_SDCARD_WRITE_Initialize(void);
 
 static bool APP_SDCARD_WIRTE_Write_SDCard(
     const DRV_HANDLE handle,
-    uint16_t* const pBuffer,
+    uint8_t* const pBuffer,
     const uint16_t bytesToWrite,
     uint16_t*const pNumBytesWrote
 );
